refactor(mc-script): compile it through babel, to support node 10
