+++ packages/mc-scripts/package.json
@@ -20,8 +20,21 @@
   "publishConfig": {
     "access": "public"
   },
+  "main": "./build/index.js",
+  "files": [
+    "bin",
+    "build",
+    "config",
+    "package.json",
+    "LICENSE",
+    "README.md"
+  ],
   "bin": {
     "mc-scripts": "./bin/mc-scripts.js"
+  },
+  "scripts": {
+    "prebuild": "rimraf build/**",
+    "build": "babel src --out-dir build"
   },
   "browserslist": {
     "production": [
@@ -36,6 +49,7 @@
   },
   "dependencies": {
     "@babel/core": "7.9.0",
+    "@babel/polyfill": "7.8.7",
     "@commercetools-frontend/assets": "16.0.0",
     "@commercetools-frontend/babel-preset-mc-app": "16.7.0",
     "@commercetools-frontend/mc-dev-authentication": "16.6.0",
@@ -79,8 +93,26 @@
     "webpack-dev-server": "3.10.3",
     "webpackbar": "4.0.0"
   },
+  "devDependencies": {
+    "@babel/preset-env": "7.9.5"
+  },
   "engines": {
-    "node": ">=8",
+    "node": ">=10",
     "npm": ">=5"
+  },
+  "babel": {
+    "presets": [
+      [
+        "@babel/preset-env",
+        {
+          "targets": {
+            "node": "10"
+          },
+          "modules": "commonjs",
+          "useBuiltIns": "usage",
+          "corejs": 3
+        }
+      ]
+    ]
   }
 }
