+++ packages/mc-scripts/src/build.js
