+++ packages/mc-scripts/src/compile-html.js
