+++ packages/mc-scripts/src/config/webpack-dev-server.config.js
