+++ packages/mc-scripts/src/extract-intl.js
