+++ packages/mc-scripts/src/start.js
