+++ packages/mc-scripts/src/webpack-plugins/final-stats-writer-plugin.js
