+++ packages/mc-scripts/src/webpack-plugins/local-html-webpack-plugin.js
